2023 day 5 puzzle 2
Time taken: ~2 hours

This was the puzzle I strugged with the most, so far.  The naïve
solution expands the full list of numbers before using the same
implementation as puzzle 1.  It is quick to implement, but takes
impossibly long to process the full input.

The proper solution compares ranges. Because input ranges can either
not overlap, or overlap partially (on the left or right), or overlap
fully, the code took a while for me to get right.  The recursive
solution first sorts the input and mapping ranges by lower bound.
The mapping range must be sorted by /source/ lower bound, not
/destination/ lower bound.  Then we recurse along both lists.  In
the case of an overlap, we have to chop up the input range, and put
the "leftovers" back on the front of the input range list when we
recurse.