2023 day 8 puzzle 2
Time taken: too much

Certain invariants that are not disclosed in the puzzle text had to
be discovered experimentally (or by searching online). With the
hidden facts, the solution is a straightforward *least common
multiple* calculation (`lcm`).

Without awareness of these invariants, the naïve algorithm takes an
impossibly long time to run.  For example, the answer for my puzzle
input was 18,215,611,419,223.  I ran the naïve solution for > 4
hours and I don't know how many iterations it did, but it sure as
heck wasn't that many!

The hidden invariants are:

1. The path from each 'A' node to the first 'Z' node has a number of
   steps that is a multiple of the (uncycled) path length.

2. The path from each of the first 'Z' nodes reaches another 'Z'
   node, and that path segment is the same length as the previous
   segment, and so on.  Incidentally, each path loops back to the
   same 'Z' node, guaranteeing this invariant in the simplest way.

Accordingly, LCM of these path lengths computes the solution.

I really do not like having to discover hidden invariants in the
input data in order to solve the problem.  This is the first such
problem (in 2023 anyway) where the problem statement is insufficient
and I hope it is the last.