2023 day 12 puzzle 2
Time taken: lots
Topics: state machine, combinatorics

Naïve implementations take "heat death of universe" time.  I found
some optimisations but after running overnight, there were still
some input lines that had not produced an answer.

The key insight was to realise that divergent paths often "collapse"
to a common state.  As a minimal example:

    .??.??. 1,1

There are 4 possible valid arrangements:

    ..#..#.  /  ..#.#..  /  .#...#.  /  .#..#..

When we reach the mid-point, there were two possible arrangements so
far.  But there is only one "state" of what remains to be done, i.e.
`.??. 1`.  After each step (consumption of exactly one character
from the template), we gather all the next possible states and
propagate each state *and its multiplicity* to the next step (a
multiset of states).  When we reach the end of the template, the sum
of the multiplicities of valid end states is the number of possible
arrangements.