2023 day 17 puzzle 1
Time taken: 3 hours
Topics: graph, shortest path

Oof, I struggled with this one.  The solution is not very efficient.
Takes about 45s to run.

While walking the graph in a kind of shortest path algorithm, we
keep track of the minimum cost per direction *and* steps in that
direction.  There is likely some low-hanging optimisation fruit.