2023 day 18 puzzle 2
Time taken: hours

Brutal.  Impossible to flood fill the closed path, so a different
approach was required.  I look at the corners of the path, line by
line, and keep track of the spans.  A simple concept, but it took me
a long time to figure out all the cases for how the spans can be
introduced, eliminated, split and merged.

One thing that helped was testing with minimal paths for particular
shapes.  Bugs and corner cases could be more easily identified with
this technique.  But it still took a long time to find all the
minimal shapes, and work through all the issues.