2023 day 18 puzzle 2 - a nicer solution
After solving it, I checked r/adventofcode and there found a much
nicer idea.  Calculate the enclosed by adding the area under the
path when moving right, and subtracting when moving left.

This is the area as to the midpoint of the path.  Then add 1/2 *
path length + 1 {- 4 * 1/4 -} to dig out the whole path.