2023 day 20 puzzle 2
Time taken: 45 minutes** / ~2 hours
Topics: combinatorics

Brute force is infeasible.  I eventually turned to manually
constructing and inspecting the parts of the graph leading to "rx".
Once you work out the "fan-in" situation and the pulses required to
send Low pulse to sink "rx", it becomes feasible to brute force the
cycle length for each of those conditions, separately.  Once you
have those values, the least-common-multiple is the answer.

** I manually found the individual conditions and it was simple to
modify the code to iterate the button pushes until each condition
arises, and output the number of steps.  Then I computed the `lcm`
(in GHCi) and collected my star.

Implementing the graph analysis in code took quite a bit longer and
ain't pretty.  But it gets the job done.