2023 day 21 puzzle 2
Time taken: hours and hours

Hoo boy.  For me this was the hardest so far.  Another one that
depends on unspoken properties of the input (which were not
reflected in the sample inputs).  And although I got quite close, a
lot of edge (or should I say "corner") cases meant I submitted a lot
of wrong guesses.  Got there in the end.

My solution is iterative yet there should be a simpler geometric
formula.  However, performance is fine, with the slowest part being
the calculation of the occupancies of the "centre" and "corner"
diamonds.

For more details see the extensive commentary in the code.